refactor(components): [checkbox]redefine the checkbox-group file path
