+++ packages/components/checkbox/src/composables/index.ts
@@ -1,2 +1,1 @@
-export * from './use-checkbox-group'
 export * from './use-checkbox'
