+++ packages/components/checkbox/src/composables/use-checkbox/index.ts
@@ -63,3 +63,5 @@ export const useCheckbox = (
     onClickRoot,
   }
 }
+
+export * from './use-group'
